+++ src/pages/_app.tsx
@@ -1,6 +1,5 @@
-import React from "react";
+import React, { useEffect } from "react";
 import type { AppProps } from "next/app";
-import gaUtils, * as gtag from "../lib/Analytics";
 import { useRouter } from "next/router";
 
 import "@nypl/design-system-react-components/dist/styles.css";
@@ -10,6 +9,8 @@ import appConfig from "~/config/appConfig";
 import { documentTitles } from "../constants/labels";
 import "@nypl/web-reader/dist/index.css";
 import { FeatureFlagProvider } from "../context/FeatureFlagContext";
+import { pageViewData } from "../constants/analytics";
+import { trackPageview } from "../lib/Analytics";
 
 /**
  * Determines if we are running on server or in the client.
@@ -19,18 +20,8 @@ function isServerRendered(): boolean {
   return typeof window === "undefined";
 }
 
-// Get the Google Analytics code for the HTML snippet below.
-const isProduction = true;
-const gaCode = gtag.getGoogleGACode(isProduction);
-// Set up Google Analytics if it isn't already. There's an HTML snippet in the
-// DOM below that initializes GA. If it fails, this tries again. The HTML
-// snippet is better since it works without javascript.
-if (!isServerRendered()) {
-  gaUtils.setupAnalytics(isProduction);
-}
-
 /**
- * Sets page title
+ * Sets page title and sends analytics data
  * @param query the router query
  * @returns the title of the page (as shown in browser tab)
  */
@@ -43,13 +34,37 @@ const setTitle = (query: any) => {
     return documentTitles.search;
   } else if (query.linkId) {
     return documentTitles.readItem;
+  } else if (query.collectionId) {
+    return documentTitles.collection;
   } else {
     return documentTitles.home;
   }
 };
 
+const sendAnalytics = (query: any) => {
+  if (query.workId) {
+    trackPageview(pageViewData.workItem);
+  } else if (query.editionId) {
+    trackPageview(pageViewData.editionItem);
+  } else if (query.query) {
+    trackPageview(pageViewData.search);
+  } else if (query.linkId) {
+    trackPageview(pageViewData.readItem);
+  } else if (query.collectionId) {
+    trackPageview(pageViewData.readItem);
+  } else {
+    trackPageview(pageViewData.home);
+  }
+};
+
 function MyApp({ Component, pageProps }: AppProps) {
   const router = useRouter();
+
+  useEffect(() => {
+    if (!isServerRendered()) {
+      sendAnalytics(router.query);
+    }
+  });
 
   return (
     <>
@@ -57,24 +72,8 @@ function MyApp({ Component, pageProps }: AppProps) {
         <title>{setTitle(router.query)}</title>
 
         <link rel="icon" href={appConfig.favIconPath} />
-        {/* <!-- Google Analytics --> */}
-        {/* We can't directly put the script into this component because React
-            doesn't allow it, so we must add it through the
-            `dangerouslySetInnerHTML` prop.
-        */}
-        <script
-          dangerouslySetInnerHTML={{
-            __html: `
-              (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-              (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-              m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-              })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-              ga('create', '${gaCode}', 'auto');
-              ga('send', 'pageview');
-            `,
-          }}
-        />
-        {/* <!-- End Google Analytics --> */}
+
+        <script src={appConfig.analytics} async />
       </Head>
       <FeatureFlagProvider>
         <Component {...pageProps} />