+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add TOC for single resource PDFs
 - Refactored Playwright/Cucumber tests to use new dropdown action function
 - Refactored Playwright/Cucumber tests to use new text fill function
+- Migrate from Google Analytics to Adobe Analytics
 
 ## [0.17.1]
 - Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app